+++ packages/analytics/src/Infra/index.ts
@@ -1,1 +1,2 @@
 export * from './Redis/RedisAnalyticsStore'
+export * from './Redis/RedisStatisticsStore'