+++ packages/domain-events-infra/src/Infra/SQS/SQSEventMessageHandler.ts
@@ -17,6 +17,8 @@ export class SQSEventMessageHandler implements DomainEventMessageHandlerInterfac
 
     const domainEvent: DomainEventInterface = JSON.parse(domainEventJson)
 
+    domainEvent.createdAt = new Date(domainEvent.createdAt)
+
     const handler = this.handlers.get(domainEvent.type)
     if (!handler) {
       this.logger.debug(`Event handler for event type ${domainEvent.type} does not exist`)