+++ packages/domain-events-infra/src/Infra/SQS/SQSNewRelicEventMessageHandler.spec.ts
@@ -32,7 +32,7 @@ describe('SQSNewRelicEventMessageHandler', () => {
 
   it('should handle messages', async () => {
     const sqsMessage = `{
-      "Message" : "eJyrViqpLEhVslIKcQ0OUdJRKkiszMlPTFGyqlZKy88HiiclFinV1gIA9tQMhA=="
+      "Message" : "eJyrViqpLEhVslIKcQ0OUdJRKkiszMlPTFGyqlZKy88HiiclFinV6iglF6UmlqSmOJYAhQwtzQ10DQyBKMTAwAqM9AwMDKOUagGlWhXt"
     }`
 
     await createHandler().handleMessage(sqsMessage)